+++ src/main/kotlin/sevenKata/JDBCUrl.kt
@@ -1,4 +1,4 @@
-package org.example
+package org.example.sevenKata
 
 class JDBCUrl {
     fun hidePasswordFromConnection( urlString: String ): String{
