+++ src/main/kotlin/sevenKata/PassGen.kt
@@ -1,4 +1,4 @@
-package org.example
+package org.example.sevenKata
 
 class PassGen {
     fun generatePassword( length: Int ): String{
