+++ src/test/kotlin/CatAndMouseTest.kt
@@ -1,4 +1,4 @@
-import org.example.CatAndMouse
+import org.example.sevenKata.CatAndMouse
 import org.junit.jupiter.api.Test
 
 import org.junit.jupiter.api.Assertions.*
