+++ src/test/kotlin/MirrorStringsTest.kt
@@ -1,4 +1,4 @@
-import org.example.MirrorStrings
+import org.example.sevenKata.MirrorStrings
 import org.junit.jupiter.api.Test
 
 import org.junit.jupiter.api.Assertions.*
