+++ src/test/kotlin/MoveTenTest.kt
@@ -1,4 +1,4 @@
-import org.example.MoveTen
+import org.example.sevenKata.MoveTen
 import org.junit.jupiter.api.Test
 
 import org.junit.jupiter.api.Assertions.*
