+++ src/test/kotlin/OfficeMoodTest.kt
@@ -1,4 +1,4 @@
-import org.example.OfficeMood
+import org.example.sevenKata.OfficeMood
 import org.junit.jupiter.api.Test
 
 import org.junit.jupiter.api.Assertions.*
