+++ src/test/kotlin/SplitStringTest.kt
@@ -1,4 +1,4 @@
-import org.example.SplitString
+import org.example.sevenKata.SplitString
 import org.junit.jupiter.api.Test
 
 import org.junit.jupiter.api.Assertions.*
